pyproject.toml
[tool.poetry]
name = "palmeiras-news"
version = "0.1.0"
description = "Web Scraping que pega informações do proximo jogo do Palmeiras no site oficial, e envia pos SMS."
authors = ["LecoOliveira <lecoverde10@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]


[tool.poetry.scripts]
palmeiras = "app.cli.palmeiras:app"
palmeiras-run = "app.main:app.run()"


[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.1"
requests = "^2.30.0"
python-dotenv = "^1.0.0"
rocketry = "^2.5.1"
typer = "^0.9.0"
rich = "^13.4.1"
cloudscraper = "^1.2.71"
python-telegram-bot = "^21.3"
pydantic = "1.10.10"
pytest-asyncio = "^0.23.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
blue = "^0.9.1"
isort = "^5.12.0"
mkdocs = "^1.4.3"
pytest-cov = "^4.0.0"
taskipy = "^1.10.4"


[tool.poetry.group.doc.dependencies]
mkdocstrings = "^0.21.2"
mkdocstrings-python = "^1.0.0"
mkdocs-bootswatch = "^1.1"
mkdocs-material = "^9.1.12"
pymdown-extensions = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"


[tool.isort]
profile = "black"
line_length = 79


[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
pre_test = "task lint"
test = "pytest -s -x --cov=. -vv"
post_test = "coverage html --omit='*/tests*' -d tests/coverage"
docs = "mkdocs serve"


[tool.coverage.run]
omit = [
    # Omite a pasta tests
    "*/tests/*",
    # Omite qualquer arquivo teste
    "app/teste.py",
]